+++ src/types.h
@@ -36,23 +36,16 @@ typedef u32 uint;
 typedef char16_t wchar;
 
 // log functions
-// only debug builds contain log because FTP has problem when the log file is not closed
 #ifdef DEBUG
 #define debug(mes) botwsavs::fs::Logger::Instance().Debug(mes)
 #define debugf(mes, value) botwsavs::fs::Logger::Instance().DebugF(mes, value)
+#else
+#define debug(mes) 
+#define debugf(mes)
+#endif
 #define info(mes) botwsavs::fs::Logger::Instance().Info(mes)
 #define warn(mes) botwsavs::fs::Logger::Instance().Warn(mes)
 #define error(mes) botwsavs::fs::Logger::Instance().Error(mes)
 #define infof(mes, value) botwsavs::fs::Logger::Instance().InfoF(mes, value)
 #define warnf(mes, value) botwsavs::fs::Logger::Instance().WarnF(mes, value)
 #define errorf(mes, value) botwsavs::fs::Logger::Instance().ErrorF(mes, value)
-#else
-#define debug(mes) 
-#define debugf(mes)
-#define info(mes)
-#define warn(mes)
-#define error(mes)
-#define infof(mes, value)
-#define warnf(mes, value)
-#define errorf(mes, value)
-#endif